server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.8.3
docx==0.2.4
